requirements.txt
# python=3.9

shapely==2.0.4
numpy==1.26.4
pillow==10.0.1
pyclipper
opencv-python <= 4.2.0.32
scikit-image==0.22.0
